+++ fvm/bootstrap.go
@@ -240,7 +240,7 @@ func deployContractTransaction(address flow.Address, contract []byte, contractNa
 		flow.NewTransactionBody().
 			SetScript([]byte(fmt.Sprintf(deployContractTransactionTemplate, contractName, hex.EncodeToString(contract)))).
 			AddAuthorizer(address),
-		uint32(0),
+		0,
 	)
 }
 
@@ -250,7 +250,7 @@ func deployFlowTokenTransaction(flowToken, service flow.Address, contract []byte
 			SetScript([]byte(fmt.Sprintf(deployFlowTokenTransactionTemplate, hex.EncodeToString(contract)))).
 			AddAuthorizer(flowToken).
 			AddAuthorizer(service),
-		uint32(0),
+		0,
 	)
 }
 
@@ -260,7 +260,7 @@ func deployFlowFeesTransaction(flowFees, service flow.Address, contract []byte) 
 			SetScript([]byte(fmt.Sprintf(deployFlowFeesTransactionTemplate, hex.EncodeToString(contract)))).
 			AddAuthorizer(flowFees).
 			AddAuthorizer(service),
-		uint32(0),
+		0,
 	)
 }
 
@@ -278,7 +278,7 @@ func mintFlowTokenTransaction(
 			SetScript([]byte(fmt.Sprintf(mintFlowTokenTransactionTemplate, fungibleToken, flowToken))).
 			AddArgument(initialSupplyArg).
 			AddAuthorizer(service),
-		uint32(0),
+		0,
 	)
 }
