+++ fvm/fees.go
@@ -64,6 +64,6 @@ func deductTransactionFeeTransaction(accountAddress, serviceAddress flow.Address
 		flow.NewTransactionBody().
 			SetScript([]byte(fmt.Sprintf(deductTransactionFeeTransactionTemplate, serviceAddress))).
 			AddAuthorizer(accountAddress),
-		uint32(0),
+		0,
 	)
 }
